+++ app/assets/javascripts/jobs/components/filtered_search/utils.js
@@ -1,0 +1,23 @@
+import { jobStatusValues } from './constants';
+
+// validates query string used for filtered search
+// on jobs table to ensure GraphQL query is called correctly
+export const validateQueryString = (queryStringObj) => {
+  // currently only one token is supported `statuses`
+  // this code will need to be expanded as more tokens
+  // are introduced
+
+  const filters = Object.keys(queryStringObj);
+
+  if (filters.includes('statuses')) {
+    const found = jobStatusValues.find((status) => status === queryStringObj.statuses);
+
+    if (found) {
+      return queryStringObj;
+    }
+
+    return null;
+  }
+
+  return null;
+};