+++ db/post_migrate/20220815061621_rename_web_hooks_service_id_to_integration_id.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class RenameWebHooksServiceIdToIntegrationId < Gitlab::Database::Migration[2.0]
+  disable_ddl_transaction!
+
+  def up
+    rename_column_concurrently :web_hooks, :service_id, :integration_id
+  end
+
+  def down
+    undo_rename_column_concurrently :web_hooks, :service_id, :integration_id
+  end
+end