+++ db/post_migrate/20220822090656_drop_build_coverage_regex_from_project.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class DropBuildCoverageRegexFromProject < Gitlab::Database::Migration[2.0]
+  enable_lock_retries!
+
+  def up
+    remove_column :projects, :build_coverage_regex
+  end
+
+  def down
+    add_column :projects, :build_coverage_regex, :string # rubocop: disable Migration/AddColumnsToWideTables
+  end
+end