+++ db/structure.sql
@@ -22,6 +22,40 @@ RETURN NULL;
 END
 $$;
 
+CREATE FUNCTION function_for_trigger_a645cee67576() RETURNS trigger
+    LANGUAGE plpgsql
+    AS $$
+BEGIN
+  NEW."service_id" := NEW."integration_id";
+  RETURN NEW;
+END
+$$;
+
+CREATE FUNCTION function_for_trigger_a87bcfdf0f0b() RETURNS trigger
+    LANGUAGE plpgsql
+    AS $$
+BEGIN
+  IF NEW."service_id" IS NULL AND NEW."integration_id" IS NOT NULL THEN
+    NEW."service_id" = NEW."integration_id";
+  END IF;
+
+  IF NEW."integration_id" IS NULL AND NEW."service_id" IS NOT NULL THEN
+    NEW."integration_id" = NEW."service_id";
+  END IF;
+
+  RETURN NEW;
+END
+$$;
+
+CREATE FUNCTION function_for_trigger_aca5c963d732() RETURNS trigger
+    LANGUAGE plpgsql
+    AS $$
+BEGIN
+  NEW."integration_id" := NEW."service_id";
+  RETURN NEW;
+END
+$$;
+
 CREATE FUNCTION gitlab_schema_prevent_write() RETURNS trigger
     LANGUAGE plpgsql
     AS $$
@@ -19981,7 +20015,6 @@ CREATE TABLE projects (
     mirror_user_id integer,
     shared_runners_enabled boolean DEFAULT true NOT NULL,
     runners_token character varying,
-    build_coverage_regex character varying,
     build_allow_git_fetch boolean DEFAULT true NOT NULL,
     build_timeout integer DEFAULT 3600 NOT NULL,
     mirror_trigger_builds boolean DEFAULT false NOT NULL,
@@ -22751,7 +22784,8 @@ CREATE TABLE web_hooks (
     backoff_count smallint DEFAULT 0 NOT NULL,
     disabled_until timestamp with time zone,
     encrypted_url_variables bytea,
-    encrypted_url_variables_iv bytea
+    encrypted_url_variables_iv bytea,
+    integration_id integer
 );
 
 CREATE SEQUENCE web_hooks_id_seq
@@ -30495,6 +30529,8 @@ CREATE INDEX index_web_hook_logs_part_on_web_hook_id ON ONLY web_hook_logs USING
 
 CREATE INDEX index_web_hooks_on_group_id ON web_hooks USING btree (group_id) WHERE ((type)::text = 'GroupHook'::text);
 
+CREATE INDEX index_web_hooks_on_integration_id ON web_hooks USING btree (integration_id);
+
 CREATE INDEX index_web_hooks_on_project_id ON web_hooks USING btree (project_id);
 
 CREATE INDEX index_web_hooks_on_project_id_recent_failures ON web_hooks USING btree (project_id, recent_failures);
@@ -31959,6 +31995,12 @@ CREATE TRIGGER nullify_merge_request_metrics_build_data_on_update BEFORE UPDATE 
 
 CREATE TRIGGER projects_loose_fk_trigger AFTER DELETE ON projects REFERENCING OLD TABLE AS old_table FOR EACH STATEMENT EXECUTE FUNCTION insert_into_loose_foreign_keys_deleted_records();
 
+CREATE TRIGGER trigger_a645cee67576 BEFORE UPDATE OF integration_id ON web_hooks FOR EACH ROW EXECUTE FUNCTION function_for_trigger_a645cee67576();
+
+CREATE TRIGGER trigger_a87bcfdf0f0b BEFORE INSERT ON web_hooks FOR EACH ROW EXECUTE FUNCTION function_for_trigger_a87bcfdf0f0b();
+
+CREATE TRIGGER trigger_aca5c963d732 BEFORE UPDATE OF service_id ON web_hooks FOR EACH ROW EXECUTE FUNCTION function_for_trigger_aca5c963d732();
+
 CREATE TRIGGER trigger_delete_project_namespace_on_project_delete AFTER DELETE ON projects FOR EACH ROW WHEN ((old.project_namespace_id IS NOT NULL)) EXECUTE FUNCTION delete_associated_project_namespace();
 
 CREATE TRIGGER trigger_has_external_issue_tracker_on_delete AFTER DELETE ON integrations FOR EACH ROW WHEN ((((old.category)::text = 'issue_tracker'::text) AND (old.active = true) AND (old.project_id IS NOT NULL))) EXECUTE FUNCTION set_has_external_issue_tracker();
@@ -32793,6 +32835,9 @@ ALTER TABLE ONLY project_group_links
 
 ALTER TABLE ONLY project_topics
     ADD CONSTRAINT fk_db13576296 FOREIGN KEY (project_id) REFERENCES projects(id) ON DELETE CASCADE;
+
+ALTER TABLE ONLY web_hooks
+    ADD CONSTRAINT fk_db1ea5699b FOREIGN KEY (integration_id) REFERENCES integrations(id) ON DELETE CASCADE;
 
 ALTER TABLE ONLY security_scans
     ADD CONSTRAINT fk_dbc89265b9 FOREIGN KEY (project_id) REFERENCES projects(id) ON DELETE CASCADE;